refactor: reorganize DTO, Command, and Validator folders in `user-service`
- Moved `DTO` files to `Application/DTO/ToCommand`.
- Relocated `Command` and `Validator` files under `Domain/Command`.
- Updated corresponding import paths across affected services and handlers.
