+++ user-service/src/Domain/Command/Handlers/BaseHandlerCommand.ts
