+++ user-service/src/Domain/Command/Handlers/DeleteUserCommandHandler.ts
